[atlas-gradle-plugin] bugfix
