+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/dependency/AtlasDependencyTree.java
@@ -333,4 +333,9 @@ public class AtlasDependencyTree {
         }
         return dependencyJson;
     }
+
+    public List<ResolvedDependencyInfo> getResolvedDependencies() {
+        return mResolvedDependencies;
+    }
+
 }
