+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/manager/AtlasLibTaskManager.java
@@ -236,8 +236,8 @@ import com.taobao.android.builder.tasks.awo.ProcessAwoAndroidResources;
 import com.taobao.android.builder.tasks.awo.maindex.DexBuildTask;
 import com.taobao.android.builder.tasks.awo.maindex.DexInstallTask;
 import com.taobao.android.builder.tasks.awo.maindex.PrepareMainDexJarsTask;
-import com.taobao.android.builder.tasks.bundle.MergeAwbAssetConfigAction;
-import com.taobao.android.builder.tasks.bundle.MergeAwbResourceConfigAction;
+import com.taobao.android.builder.tasks.app.merge.bundle.MergeAwbAssetConfigAction;
+import com.taobao.android.builder.tasks.app.merge.bundle.MergeAwbResourceConfigAction;
 import com.taobao.android.builder.tasks.library.AwbGenerator;
 import com.taobao.android.builder.tasks.library.JarExtractTask;
 import com.taobao.android.builder.tasks.library.publish.UpdatePomTask;
