+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/tasks/app/LogDependenciesTask.java
@@ -381,7 +381,7 @@ public class LogDependenciesTask extends BaseTask {
 
         @Override
         public String getName() {
-            return scope.getTaskName("log", "dependencies");
+            return scope.getTaskName("log", "Dependencies");
         }
 
         @Override
