+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/tasks/app/databinding/AwbDataBindingExportBuildInfoTask.java
@@ -228,7 +228,7 @@ import com.google.common.collect.Iterables;
 import com.taobao.android.builder.AtlasBuildContext;
 import com.taobao.android.builder.dependency.AtlasDependencyTree;
 import com.taobao.android.builder.dependency.model.AwbBundle;
-import com.taobao.android.builder.tasks.app.bundle.TaskCreater;
+import com.taobao.android.builder.tasks.manager.TaskCreater;
 import com.taobao.android.builder.tasks.manager.MtlBaseTaskAction;
 import com.taobao.android.builder.tools.concurrent.ExecutorServicesHelper;
 import org.gradle.api.file.ConfigurableFileTree;
