--- atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/tasks/app/merge/MergeAwbManifests.java
@@ -1,598 +1,0 @@
-/*
- *
- *
- *
- *                                   Apache License
- *                             Version 2.0, January 2004
- *                          http://www.apache.org/licenses/
- *
- *     TERMS AND CONDITIONS FOR USE, REPRODUCTION, AND DISTRIBUTION
- *
- *     1. Definitions.
- *
- *        "License" shall mean the terms and conditions for use, reproduction,
- *        and distribution as defined by Sections 1 through 9 of this document.
- *
- *        "Licensor" shall mean the copyright owner or entity authorized by
- *        the copyright owner that is granting the License.
- *
- *        "Legal Entity" shall mean the union of the acting entity and all
- *        other entities that control, are controlled by, or are under common
- *        control with that entity. For the purposes of this definition,
- *        "control" means (i) the power, direct or indirect, to cause the
- *        direction or management of such entity, whether by contract or
- *        otherwise, or (ii) ownership of fifty percent (50%) or more of the
- *        outstanding shares, or (iii) beneficial ownership of such entity.
- *
- *        "You" (or "Your") shall mean an individual or Legal Entity
- *        exercising permissions granted by this License.
- *
- *        "Source" form shall mean the preferred form for making modifications,
- *        including but not limited to software source code, documentation
- *        source, and configuration files.
- *
- *        "Object" form shall mean any form resulting from mechanical
- *        transformation or translation of a Source form, including but
- *        not limited to compiled object code, generated documentation,
- *        and conversions to other media types.
- *
- *        "Work" shall mean the work of authorship, whether in Source or
- *        Object form, made available under the License, as indicated by a
- *        copyright notice that is included in or attached to the work
- *        (an example is provided in the Appendix below).
- *
- *        "Derivative Works" shall mean any work, whether in Source or Object
- *        form, that is based on (or derived from) the Work and for which the
- *        editorial revisions, annotations, elaborations, or other modifications
- *        represent, as a whole, an original work of authorship. For the purposes
- *        of this License, Derivative Works shall not include works that remain
- *        separable from, or merely link (or bind by name) to the interfaces of,
- *        the Work and Derivative Works thereof.
- *
- *        "Contribution" shall mean any work of authorship, including
- *        the original version of the Work and any modifications or additions
- *        to that Work or Derivative Works thereof, that is intentionally
- *        submitted to Licensor for inclusion in the Work by the copyright owner
- *        or by an individual or Legal Entity authorized to submit on behalf of
- *        the copyright owner. For the purposes of this definition, "submitted"
- *        means any form of electronic, verbal, or written communication sent
- *        to the Licensor or its representatives, including but not limited to
- *        communication on electronic mailing lists, source code control systems,
- *        and issue tracking systems that are managed by, or on behalf of, the
- *        Licensor for the purpose of discussing and improving the Work, but
- *        excluding communication that is conspicuously marked or otherwise
- *        designated in writing by the copyright owner as "Not a Contribution."
- *
- *        "Contributor" shall mean Licensor and any individual or Legal Entity
- *        on behalf of whom a Contribution has been received by Licensor and
- *        subsequently incorporated within the Work.
- *
- *     2. Grant of Copyright License. Subject to the terms and conditions of
- *        this License, each Contributor hereby grants to You a perpetual,
- *        worldwide, non-exclusive, no-charge, royalty-free, irrevocable
- *        copyright license to reproduce, prepare Derivative Works of,
- *        publicly display, publicly perform, sublicense, and distribute the
- *        Work and such Derivative Works in Source or Object form.
- *
- *     3. Grant of Patent License. Subject to the terms and conditions of
- *        this License, each Contributor hereby grants to You a perpetual,
- *        worldwide, non-exclusive, no-charge, royalty-free, irrevocable
- *        (except as stated in this section) patent license to make, have made,
- *        use, offer to sell, sell, import, and otherwise transfer the Work,
- *        where such license applies only to those patent claims licensable
- *        by such Contributor that are necessarily infringed by their
- *        Contribution(s) alone or by combination of their Contribution(s)
- *        with the Work to which such Contribution(s) was submitted. If You
- *        institute patent litigation against any entity (including a
- *        cross-claim or counterclaim in a lawsuit) alleging that the Work
- *        or a Contribution incorporated within the Work constitutes direct
- *        or contributory patent infringement, then any patent licenses
- *        granted to You under this License for that Work shall terminate
- *        as of the date such litigation is filed.
- *
- *     4. Redistribution. You may reproduce and distribute copies of the
- *        Work or Derivative Works thereof in any medium, with or without
- *        modifications, and in Source or Object form, provided that You
- *        meet the following conditions:
- *
- *        (a) You must give any other recipients of the Work or
- *            Derivative Works a copy of this License; and
- *
- *        (b) You must cause any modified files to carry prominent notices
- *            stating that You changed the files; and
- *
- *        (c) You must retain, in the Source form of any Derivative Works
- *            that You distribute, all copyright, patent, trademark, and
- *            attribution notices from the Source form of the Work,
- *            excluding those notices that do not pertain to any part of
- *            the Derivative Works; and
- *
- *        (d) If the Work includes a "NOTICE" text file as part of its
- *            distribution, then any Derivative Works that You distribute must
- *            include a readable copy of the attribution notices contained
- *            within such NOTICE file, excluding those notices that do not
- *            pertain to any part of the Derivative Works, in at least one
- *            of the following places: within a NOTICE text file distributed
- *            as part of the Derivative Works; within the Source form or
- *            documentation, if provided along with the Derivative Works; or,
- *            within a display generated by the Derivative Works, if and
- *            wherever such third-party notices normally appear. The contents
- *            of the NOTICE file are for informational purposes only and
- *            do not modify the License. You may add Your own attribution
- *            notices within Derivative Works that You distribute, alongside
- *            or as an addendum to the NOTICE text from the Work, provided
- *            that such additional attribution notices cannot be construed
- *            as modifying the License.
- *
- *        You may add Your own copyright statement to Your modifications and
- *        may provide additional or different license terms and conditions
- *        for use, reproduction, or distribution of Your modifications, or
- *        for any such Derivative Works as a whole, provided Your use,
- *        reproduction, and distribution of the Work otherwise complies with
- *        the conditions stated in this License.
- *
- *     5. Submission of Contributions. Unless You explicitly state otherwise,
- *        any Contribution intentionally submitted for inclusion in the Work
- *        by You to the Licensor shall be under the terms and conditions of
- *        this License, without any additional terms or conditions.
- *        Notwithstanding the above, nothing herein shall supersede or modify
- *        the terms of any separate license agreement you may have executed
- *        with Licensor regarding such Contributions.
- *
- *     6. Trademarks. This License does not grant permission to use the trade
- *        names, trademarks, service marks, or product names of the Licensor,
- *        except as required for reasonable and customary use in describing the
- *        origin of the Work and reproducing the content of the NOTICE file.
- *
- *     7. Disclaimer of Warranty. Unless required by applicable law or
- *        agreed to in writing, Licensor provides the Work (and each
- *        Contributor provides its Contributions) on an "AS IS" BASIS,
- *        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- *        implied, including, without limitation, any warranties or conditions
- *        of TITLE, NON-INFRINGEMENT, MERCHANTABILITY, or FITNESS FOR A
- *        PARTICULAR PURPOSE. You are solely responsible for determining the
- *        appropriateness of using or redistributing the Work and assume any
- *        risks associated with Your exercise of permissions under this License.
- *
- *     8. Limitation of Liability. In no event and under no legal theory,
- *        whether in tort (including negligence), contract, or otherwise,
- *        unless required by applicable law (such as deliberate and grossly
- *        negligent acts) or agreed to in writing, shall any Contributor be
- *        liable to You for damages, including any direct, indirect, special,
- *        incidental, or consequential damages of any character arising as a
- *        result of this License or out of the use or inability to use the
- *        Work (including but not limited to damages for loss of goodwill,
- *        work stoppage, computer failure or malfunction, or any and all
- *        other commercial damages or losses), even if such Contributor
- *        has been advised of the possibility of such damages.
- *
- *     9. Accepting Warranty or Additional Liability. While redistributing
- *        the Work or Derivative Works thereof, You may choose to offer,
- *        and charge a fee for, acceptance of support, warranty, indemnity,
- *        or other liability obligations and/or rights consistent with this
- *        License. However, in accepting such obligations, You may act only
- *        on Your own behalf and on Your sole responsibility, not on behalf
- *        of any other Contributor, and only if You agree to indemnify,
- *        defend, and hold each Contributor harmless for any liability
- *        incurred by, or claims asserted against, such Contributor by reason
- *        of your accepting any such warranty or additional liability.
- *
- *     END OF TERMS AND CONDITIONS
- *
- *     APPENDIX: How to apply the Apache License to your work.
- *
- *        To apply the Apache License to your work, attach the following
- *        boilerplate notice, with the fields enclosed by brackets "[]"
- *        replaced with your own identifying information. (Don't include
- *        the brackets!)  The text should be enclosed in the appropriate
- *        comment syntax for the file format. We also recommend that a
- *        file or class name and description of purpose be included on the
- *        same "printed page" as the copyright notice for easier
- *        identification within third-party archives.
- *
- *     Copyright 2016 Alibaba Group
- *
- *     Licensed under the Apache License, Version 2.0 (the "License");
- *     you may not use this file except in compliance with the License.
- *     You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- *     Unless required by applicable law or agreed to in writing, software
- *     distributed under the License is distributed on an "AS IS" BASIS,
- *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *     See the License for the specific language governing permissions and
- *     limitations under the License.
- *
- *
- */
-package com.taobao.android.builder.tasks.app.merge;
-
-import com.android.annotations.NonNull;
-import com.android.build.gradle.internal.api.AppVariantOutputContext;
-import com.android.build.gradle.internal.dsl.CoreBuildType;
-import com.android.build.gradle.internal.dsl.CoreProductFlavor;
-import com.android.build.gradle.internal.scope.ConventionMappingHelper;
-import com.android.build.gradle.internal.scope.TaskConfigAction;
-import com.android.build.gradle.internal.scope.VariantOutputScope;
-import com.android.build.gradle.internal.variant.ApkVariantOutputData;
-import com.android.build.gradle.internal.variant.BaseVariantData;
-import com.android.build.gradle.internal.variant.BaseVariantOutputData;
-import com.android.build.gradle.tasks.ManifestProcessorTask;
-import com.android.builder.core.VariantConfiguration;
-import com.android.builder.model.AndroidLibrary;
-import com.android.builder.model.ApiVersion;
-import com.android.manifmerger.ManifestMerger2;
-import com.android.manifmerger.ManifestMerger2.Invoker.Feature;
-import com.android.manifmerger.ManifestProvider;
-import com.android.utils.FileUtils;
-import com.google.common.collect.Lists;
-import com.taobao.android.builder.dependency.model.AwbBundle;
-import com.taobao.android.builder.tools.manifest.ManifestHelper;
-
-import org.gradle.api.tasks.Input;
-import org.gradle.api.tasks.InputFile;
-import org.gradle.api.tasks.InputFiles;
-import org.gradle.api.tasks.Optional;
-import org.gradle.api.tasks.OutputFile;
-import org.gradle.api.tasks.ParallelizableTask;
-
-import java.io.File;
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.Callable;
-
-/**
- * A task that processes the manifest
- */
-@ParallelizableTask
-public class MergeAwbManifests extends ManifestProcessorTask {
-
-    private String minSdkVersion;
-
-    private String targetSdkVersion;
-
-    private Integer maxSdkVersion;
-
-    private File reportFile;
-
-    private VariantConfiguration<CoreBuildType, CoreProductFlavor, CoreProductFlavor> variantConfiguration;
-
-    private ApkVariantOutputData variantOutputData;
-
-    private List<ManifestProvider> libraries;
-
-    private List<Feature> optionalFeatures;
-
-    private File mainManifest;
-
-    private AwbBundle awbBundle;
-
-    @Override
-    protected void doFullTaskAction() {
-        getBuilder().mergeManifestsForApplication(getMainManifest(),
-                                                  getManifestOverlays(),
-                                                  getLibraries(),
-                                                  getPackageOverride(),
-                                                  getVersionCode(),
-                                                  getVersionName(),
-                                                  getMinSdkVersion(),
-                                                  getTargetSdkVersion(),
-                                                  getMaxSdkVersion(),
-                                                  getManifestOutputFile().getAbsolutePath(),
-                                                  // no aapt friendly merged manifest file necessary for applications.
-                                                  null /* aaptFriendlyManifestOutputFile */,
-                                                  getInstantRunManifestOutputFile().getAbsolutePath(),
-                                                  ManifestMerger2.MergeType.APPLICATION,
-                                                  variantConfiguration.getManifestPlaceholders(),
-                                                  getOptionalFeatures(),
-                                                  getReportFile());
-    }
-
-    @InputFile
-    public File getMainManifest() {
-        return mainManifest;
-    }
-
-    public void setMainManifest(File mainManifest) {
-        this.mainManifest = mainManifest;
-    }
-
-    @InputFiles
-    public List<File> getManifestOverlays() {
-        return variantConfiguration.getManifestOverlays();
-    }
-
-    @Input
-    @Optional
-    public String getPackageOverride() {
-        //不能返回和manbundle一样的package,不然aapt在merge resource 的时候会报资源冲突
-        //return variantConfiguration.getIdOverride();
-        return null;
-    }
-
-    @Input
-    public int getVersionCode() {
-        if (variantOutputData != null) {
-            return variantOutputData.getVersionCode();
-        }
-        return variantConfiguration.getVersionCode();
-    }
-
-    @Input
-    @Optional
-    public String getVersionName() {
-        return awbBundle.getResolvedCoordinates().getVersion();
-        //        if (variantOutputData != null) {
-        //            return variantOutputData.getVersionName();
-        //        }
-        //        return variantConfiguration.getVersionName();
-    }
-
-    /**
-     * Returns a serialized version of our map of key value pairs for placeholder substitution.
-     * <p>
-     * This serialized form is only used by gradle to compare past and present tasks to determine
-     * whether a task need to be re-run or not.
-     */
-    @SuppressWarnings("unused")
-    @Input
-    @Optional
-    public String getManifestPlaceholders() {
-        return serializeMap(variantConfiguration.getManifestPlaceholders());
-    }
-
-    /**
-     * A synthetic input to allow gradle up-to-date checks to work.
-     * <p>
-     * Since List<ManifestDependencyImpl> can't be used directly, as @Nested doesn't work on lists,
-     * this method gathers and returns the underlying manifest files.
-     */
-    @SuppressWarnings("unused")
-    @InputFiles
-    List<File> getLibraryManifests() {
-        List<ManifestProvider> libs = getLibraries();
-        if (libs == null || libs.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        List<File> files = Lists.newArrayListWithCapacity(libs.size() * 2);
-        for (ManifestProvider mdi : libs) {
-            files.add(mdi.getManifest());
-        }
-
-        return files;
-    }
-
-    @Input
-    @Optional
-    public String getMinSdkVersion() {
-        return minSdkVersion;
-    }
-
-    public void setMinSdkVersion(String minSdkVersion) {
-        this.minSdkVersion = minSdkVersion;
-    }
-
-    @Input
-    @Optional
-    public String getTargetSdkVersion() {
-        return targetSdkVersion;
-    }
-
-    public void setTargetSdkVersion(String targetSdkVersion) {
-        this.targetSdkVersion = targetSdkVersion;
-    }
-
-    @Input
-    @Optional
-    public Integer getMaxSdkVersion() {
-        return maxSdkVersion;
-    }
-
-    public void setMaxSdkVersion(Integer maxSdkVersion) {
-        this.maxSdkVersion = maxSdkVersion;
-    }
-
-    @OutputFile
-    @Optional
-    public File getReportFile() {
-        return reportFile;
-    }
-
-    public void setReportFile(File reportFile) {
-        this.reportFile = reportFile;
-    }
-
-    @Input
-    public List<Feature> getOptionalFeatures() {
-        return optionalFeatures;
-    }
-
-    public void setVariantConfiguration(VariantConfiguration<CoreBuildType, CoreProductFlavor, CoreProductFlavor> variantConfiguration) {
-        this.variantConfiguration = variantConfiguration;
-    }
-
-    public ApkVariantOutputData getVariantOutputData() {
-        return variantOutputData;
-    }
-
-    public void setVariantOutputData(ApkVariantOutputData variantOutputData) {
-        this.variantOutputData = variantOutputData;
-    }
-
-    public List<ManifestProvider> getLibraries() {
-        return libraries;
-    }
-
-    public void setLibraries(List<ManifestProvider> libraries) {
-        this.libraries = libraries;
-    }
-
-    public static class ConfigAction implements TaskConfigAction<MergeAwbManifests> {
-
-        private final VariantOutputScope scope;
-
-        private final List<Feature> optionalFeatures;
-
-        private final AwbBundle awbBundle;
-
-        private final AppVariantOutputContext appVariantOutputContext;
-
-        public ConfigAction(VariantOutputScope scope,
-                            AwbBundle awbBundle,
-                            List<Feature> optionalFeatures,
-                            AppVariantOutputContext appVariantOutputContext) {
-            this.scope = scope;
-            this.awbBundle = awbBundle;
-            this.optionalFeatures = optionalFeatures;
-            this.appVariantOutputContext = appVariantOutputContext;
-        }
-
-        @NonNull
-        @Override
-        public String getName() {
-            return scope.getTaskName("process", "AwbManifest[" + awbBundle.getName() + "]");
-        }
-
-        @NonNull
-        @Override
-        public Class<MergeAwbManifests> getType() {
-            return MergeAwbManifests.class;
-        }
-
-        @Override
-        public void execute(@NonNull MergeAwbManifests processManifestTask) {
-            BaseVariantOutputData variantOutputData = scope.getVariantOutputData();
-
-            final BaseVariantData<? extends BaseVariantOutputData> variantData = scope.getVariantScope()
-                    .getVariantData();
-            final VariantConfiguration<CoreBuildType, CoreProductFlavor, CoreProductFlavor> config = variantData
-                    .getVariantConfiguration();
-
-            processManifestTask.setAndroidBuilder(scope.getGlobalScope().getAndroidBuilder());
-            processManifestTask.setVariantName(config.getFullName());
-
-            processManifestTask.setVariantConfiguration(config);
-            if (variantOutputData instanceof ApkVariantOutputData) {
-                processManifestTask.variantOutputData = (ApkVariantOutputData) variantOutputData;
-            }
-
-            //            processManifestTask.setMainManifest(awbBundle.getManifest());
-            processManifestTask.awbBundle = awbBundle;
-
-            ConventionMappingHelper.map(processManifestTask, "mainManifest", new Callable<File>() {
-                @Override
-                public File call() throws Exception {
-                    return ManifestHelper.getOrgManifestFile(awbBundle.getAndroidLibrary());
-                }
-            });
-
-            ConventionMappingHelper.map(processManifestTask,
-                                        "libraries",
-                                        new Callable<List<ManifestProvider>>() {
-                                            @Override
-                                            public List<ManifestProvider> call() throws Exception {
-                                               return  awbBundle.getManifestProviders();
-                                            }
-                                        });
-
-            ConventionMappingHelper.map(processManifestTask,
-                                        "minSdkVersion",
-                                        new Callable<String>() {
-                                            @Override
-                                            public String call() throws Exception {
-                                                if (scope.getGlobalScope()
-                                                        .getAndroidBuilder()
-                                                        .isPreviewTarget()) {
-                                                    return scope.getGlobalScope()
-                                                            .getAndroidBuilder()
-                                                            .getTargetCodename();
-                                                }
-
-                                                ApiVersion minSdk = config.getMergedFlavor()
-                                                        .getMinSdkVersion();
-                                                return minSdk ==
-                                                        null ? null : minSdk.getApiString();
-                                            }
-                                        });
-
-            ConventionMappingHelper.map(processManifestTask,
-                                        "targetSdkVersion",
-                                        new Callable<String>() {
-                                            @Override
-                                            public String call() throws Exception {
-                                                if (scope.getGlobalScope()
-                                                        .getAndroidBuilder()
-                                                        .isPreviewTarget()) {
-                                                    return scope.getGlobalScope()
-                                                            .getAndroidBuilder()
-                                                            .getTargetCodename();
-                                                }
-                                                ApiVersion targetSdk = config.getMergedFlavor()
-                                                        .getTargetSdkVersion();
-                                                return targetSdk ==
-                                                        null ? null : targetSdk.getApiString();
-                                            }
-                                        });
-
-            ConventionMappingHelper.map(processManifestTask,
-                                        "maxSdkVersion",
-                                        new Callable<Integer>() {
-                                            @Override
-                                            public Integer call() throws Exception {
-                                                if (scope.getGlobalScope()
-                                                        .getAndroidBuilder()
-                                                        .isPreviewTarget()) {
-                                                    return null;
-                                                }
-                                                return config.getMergedFlavor().getMaxSdkVersion();
-                                            }
-                                        });
-
-            File manifestOutFile = new File(scope.getGlobalScope().getIntermediatesDir(),
-                                            "/manifests-awb/full/" +
-                                                    variantOutputData.getDirName() +
-                                                    "/" +
-                                                    awbBundle.getName() +
-                                                    "/AndroidManifest.xml");
-            processManifestTask.setManifestOutputFile(manifestOutFile);
-
-            File instantRunManifestOutputFile = FileUtils.join(scope.getGlobalScope()
-                                                                       .getIntermediatesDir(),
-                                                               "bundles-awb",
-                                                               variantOutputData.getDirName(),
-                                                               "instant-run",
-                                                               awbBundle.getName(),
-                                                               "AndroidManifest.xml");
-            processManifestTask.setInstantRunManifestOutputFile(instantRunManifestOutputFile);
-            File reportFile = FileUtils.join(scope.getGlobalScope().getOutputsDir(),
-                                             "logs",
-                                             "manifest-merger-" +
-                                                     variantData.getVariantConfiguration()
-                                                             .getBaseName() +
-                                                     "-" +
-                                                     awbBundle.getName() +
-                                                     "-report.txt");
-            processManifestTask.setReportFile(reportFile);
-            processManifestTask.optionalFeatures = optionalFeatures;
-            appVariantOutputContext.getMergeAwbManifestsMap()
-                    .put(awbBundle.getName(), processManifestTask);
-        }
-
-        @NonNull
-        private static List<? extends AndroidLibrary> getManifestDependencies(List<? extends AndroidLibrary> libraries) {
-
-            List list = Lists.newArrayListWithCapacity(libraries.size());
-
-            for (AndroidLibrary lib : libraries) {
-                if (!lib.isOptional()) {
-                    // get the dependencies
-                    list.add(lib);
-                }
-            }
-
-            return list;
-        }
-    }
-}
